Enhance `Piece` resource and related functionalities: refine validation rules, update models and migrations for nullable POV and tense fields, and add default status to `piece_statuses`. Update `GeneratorAiService` for logging and parameter naming consistency. Upgrade dependencies and adjust Horizon configuration.

+++ api/database/migrations/2025_07_02_182452_add_fields_to_piece_types.php
@@ -13,11 +13,11 @@ return new class extends Migration
                 ->after('name')
                 ->nullable();
 
-            $table->unsignedSmallInteger('min_count')
+            $table->unsignedInteger('min_count')
                 ->after('description')
                 ->default(0);
 
-            $table->unsignedSmallInteger('max_count')
+            $table->unsignedInteger('max_count')
                 ->after('min_count')
                 ->default(0);
